Move and forward basic packages to v2
- flag
- label
- merger
- pathtools
- rule
- testools

These packages won't require substantial changes in v2, so we don't need
two separate sets of definitions. v2 packages should eventually not
import anything from v1, so v2 is the source of truth. Consequently,
the v1 packages are now just aliases for v2.

This PR was mostly generated with a small program,
https://github.com/jayconrod/bazel-gazelle/blob/v2/tools/forward/forward.go.
It's throw-away code, so I'm not planning to check it in.

For #2272
