+++ internal/go_repository_tools_srcs.bzl
@@ -140,9 +140,31 @@ GO_REPOSITORY_TOOLS_SRCS = [
     Label("//v2/cmd/gazelle/update:print.go"),
     Label("//v2/cmd/gazelle/update:profiler.go"),
     Label("//v2/cmd/gazelle/update:update.go"),
+    Label("//v2/flag:BUILD.bazel"),
+    Label("//v2/flag:flag.go"),
     Label("//v2/internal:BUILD.bazel"),
     Label("//v2/internal/wspace:BUILD.bazel"),
     Label("//v2/internal/wspace:finder.go"),
+    Label("//v2/label:BUILD.bazel"),
+    Label("//v2/label:label.go"),
+    Label("//v2/merger:BUILD.bazel"),
+    Label("//v2/merger:fix.go"),
+    Label("//v2/merger:merger.go"),
+    Label("//v2/pathtools:BUILD.bazel"),
+    Label("//v2/pathtools:path.go"),
+    Label("//v2/rule:BUILD.bazel"),
+    Label("//v2/rule:directives.go"),
+    Label("//v2/rule:expr.go"),
+    Label("//v2/rule:merge.go"),
+    Label("//v2/rule:platform.go"),
+    Label("//v2/rule:platform_strings.go"),
+    Label("//v2/rule:rule.go"),
+    Label("//v2/rule:sort_labels.go"),
+    Label("//v2/rule:types.go"),
+    Label("//v2/rule:value.go"),
+    Label("//v2/testtools:BUILD.bazel"),
+    Label("//v2/testtools:config.go"),
+    Label("//v2/testtools:files.go"),
     Label("//walk:BUILD.bazel"),
     Label("//walk:cache.go"),
     Label("//walk:config.go"),
