+++ label/BUILD.bazel
@@ -1,20 +1,11 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "label",
     srcs = ["label.go"],
     importpath = "github.com/bazelbuild/bazel-gazelle/label",
     visibility = ["//visibility:public"],
-    deps = [
-        "//pathtools",
-        "@com_github_bazelbuild_buildtools//build",
-    ],
-)
-
-go_test(
-    name = "label_test",
-    srcs = ["label_test.go"],
-    embed = [":label"],
+    deps = ["//v2/label"],
 )
 
 filegroup(
@@ -23,7 +14,6 @@ filegroup(
     srcs = [
         "BUILD.bazel",
         "label.go",
-        "label_test.go",
     ],
     visibility = ["//visibility:public"],
 )
