+++ rule/types.go
@@ -15,42 +15,15 @@ limitations under the License.
 
 package rule
 
+import v2 "github.com/bazel-contrib/bazel-gazelle/v2/rule"
+
 // LoadInfo describes a file that Gazelle knows about and the symbols
 // it defines.
-type LoadInfo struct {
-	Name    string
-	Symbols []string
-	After   []string
-}
+//
+// Deprecated: Use github.com/bazel-contrib/bazel-gazelle/v2/rule.LoadInfo instead.
+type LoadInfo = v2.LoadInfo
 
 // KindInfo stores metadata for a kind of rule, for example, "go_library".
-type KindInfo struct {
-	// MatchAny is true if a rule of this kind may be matched with any rule
-	// of the same kind, regardless of attributes, if exactly one rule is
-	// present a build file.
-	MatchAny bool
-
-	// MatchAttrs is a list of attributes used in matching. For example,
-	// for go_library, this list contains "importpath". Attributes are matched
-	// in order.
-	MatchAttrs []string
-
-	// NonEmptyAttrs is a set of attributes that, if present, disqualify a rule
-	// from being deleted after merge.
-	NonEmptyAttrs map[string]bool
-
-	// SubstituteAttrs is a set of attributes that should be substituted
-	// after matching and before merging. For example, suppose generated rule A
-	// references B via an "embed" attribute, and B matches against rule C.
-	// The label for B in A's "embed" must be substituted with a label for C.
-	// "embed" would need to be in this set.
-	SubstituteAttrs map[string]bool
-
-	// MergeableAttrs is a set of attributes that should be merged before
-	// dependency resolution. See rule.Merge.
-	MergeableAttrs map[string]bool
-
-	// ResolveAttrs is a set of attributes that should be merged after
-	// dependency resolution. See rule.Merge.
-	ResolveAttrs map[string]bool
-}
+//
+// Deprecated: Use github.com/bazel-contrib/bazel-gazelle/v2/rule.KindInfo instead.
+type KindInfo = v2.KindInfo
