+++ v2/cmd/gazelle/update/BUILD.bazel
@@ -19,14 +19,14 @@ go_library(
     },
     deps = [
         "//config",
-        "//flag",
-        "//label",
         "//language",
-        "//merger",
         "//repo",
         "//resolve",
-        "//rule",
+        "//v2/flag",
         "//v2/internal/wspace",
+        "//v2/label",
+        "//v2/merger",
+        "//v2/rule",
         "//walk",
         "@com_github_bazelbuild_buildtools//build",
         "@com_github_pmezard_go_difflib//difflib",
