+++ v2/cmd/gazelle/update/diff.go
@@ -22,8 +22,8 @@ import (
 	"os"
 	"path/filepath"
 
+	"github.com/bazel-contrib/bazel-gazelle/v2/rule"
 	"github.com/bazelbuild/bazel-gazelle/config"
-	"github.com/bazelbuild/bazel-gazelle/rule"
 	"github.com/pmezard/go-difflib/difflib"
 )
