+++ v2/cmd/gazelle/update/fix.go
@@ -21,8 +21,8 @@ import (
 	"os"
 	"path/filepath"
 
+	"github.com/bazel-contrib/bazel-gazelle/v2/rule"
 	"github.com/bazelbuild/bazel-gazelle/config"
-	"github.com/bazelbuild/bazel-gazelle/rule"
 )
 
 func fixFile(c *config.Config, f *rule.File) error {
