+++ v2/cmd/gazelle/update/print.go
@@ -19,8 +19,8 @@ import (
 	"fmt"
 	"os"
 
+	"github.com/bazel-contrib/bazel-gazelle/v2/rule"
 	"github.com/bazelbuild/bazel-gazelle/config"
-	"github.com/bazelbuild/bazel-gazelle/rule"
 )
 
 func printFile(c *config.Config, f *rule.File) error {
