+++ v2/flag/BUILD.bazel
@@ -3,9 +3,8 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "flag",
     srcs = ["flag.go"],
-    importpath = "github.com/bazelbuild/bazel-gazelle/flag",
+    importpath = "github.com/bazel-contrib/bazel-gazelle/v2/flag",
     visibility = ["//visibility:public"],
-    deps = ["//v2/flag"],
 )
 
 filegroup(
@@ -15,11 +14,5 @@ filegroup(
         "BUILD.bazel",
         "flag.go",
     ],
-    visibility = ["//visibility:public"],
-)
-
-alias(
-    name = "go_default_library",
-    actual = ":flag",
     visibility = ["//visibility:public"],
 )
