+++ v2/merger/fix_test.go
@@ -19,8 +19,8 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/bazelbuild/bazel-gazelle/merger"
-	"github.com/bazelbuild/bazel-gazelle/rule"
+	"github.com/bazel-contrib/bazel-gazelle/v2/merger"
+	"github.com/bazel-contrib/bazel-gazelle/v2/rule"
 	bzl "github.com/bazelbuild/buildtools/build"
 	"github.com/google/go-cmp/cmp"
 )
@@ -257,12 +257,12 @@ foo_binary(
 `,
 		},
 		"assignment expr rhs list": {
-    		input: `foo_binary(
+			input: `foo_binary(
     name = "a",
     field = [magic_macro],
 )
 `,
-    		want: `load("@bar", "magic_macro")
+			want: `load("@bar", "magic_macro")
 load("@foo", "foo_binary")
 
 foo_binary(
