+++ v2/pathtools/BUILD.bazel
@@ -1,0 +1,26 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+
+go_library(
+    name = "pathtools",
+    srcs = ["path.go"],
+    importpath = "github.com/bazel-contrib/bazel-gazelle/v2/pathtools",
+    visibility = ["//visibility:public"],
+)
+
+go_test(
+    name = "pathtools_test",
+    srcs = ["path_test.go"],
+    embed = [":pathtools"],
+    deps = ["@com_github_google_go_cmp//cmp"],
+)
+
+filegroup(
+    name = "all_files",
+    testonly = True,
+    srcs = [
+        "BUILD.bazel",
+        "path.go",
+        "path_test.go",
+    ],
+    visibility = ["//visibility:public"],
+)
