+++ v2/pathtools/path_test.go
