+++ v2/rule/BUILD.bazel
@@ -1,4 +1,4 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "rule",
@@ -13,11 +13,28 @@ go_library(
         "types.go",
         "value.go",
     ],
-    importpath = "github.com/bazelbuild/bazel-gazelle/rule",
+    importpath = "github.com/bazel-contrib/bazel-gazelle/v2/rule",
     visibility = ["//visibility:public"],
     deps = [
-        "//v2/rule",
+        "//label",
         "@com_github_bazelbuild_buildtools//build",
+        "@com_github_bazelbuild_buildtools//tables",
+    ],
+)
+
+go_test(
+    name = "rule_test",
+    srcs = [
+        "directives_test.go",
+        "merge_test.go",
+        "rule_test.go",
+        "value_test.go",
+    ],
+    embed = [":rule"],
+    deps = [
+        "//label",
+        "@com_github_bazelbuild_buildtools//build",
+        "@com_github_google_go_cmp//cmp",
     ],
 )
 
@@ -27,20 +44,18 @@ filegroup(
     srcs = [
         "BUILD.bazel",
         "directives.go",
+        "directives_test.go",
         "expr.go",
         "merge.go",
+        "merge_test.go",
         "platform.go",
         "platform_strings.go",
         "rule.go",
+        "rule_test.go",
         "sort_labels.go",
         "types.go",
         "value.go",
+        "value_test.go",
     ],
-    visibility = ["//visibility:public"],
-)
-
-alias(
-    name = "go_default_library",
-    actual = ":rule",
     visibility = ["//visibility:public"],
 )
