+++ v2/rule/directives_test.go
