+++ v2/rule/merge_test.go
@@ -18,7 +18,7 @@ package rule_test
 import (
 	"testing"
 
-	"github.com/bazelbuild/bazel-gazelle/rule"
+	"github.com/bazel-contrib/bazel-gazelle/v2/rule"
 	bzl "github.com/bazelbuild/buildtools/build"
 )
