+++ v2/rule/value_test.go
